Create New Baseline File Structure
