+++ custom_components/teamtracker/clear_values.py
@@ -1,0 +1,71 @@
+async def async_clear_values() -> dict:
+    """Clear all state attributes"""
+    new_values = {}
+
+    # Reset values
+    new_values = {
+        "sport": None,
+        "league": None,
+        "league_logo": None,
+        "team_abbr": None,
+        "opponent_abbr": None,
+
+        "event_name": None,
+        "date": None,
+        "kickoff_in": None,
+        "venue": None,
+        "location": None,
+        "tv_network": None,
+        "odds": None,
+        "overunder": None,
+
+        "team_name": None,
+        "team_id": None,
+        "team_record": None,
+        "team_rank": None,
+        "team_homeaway": None,
+        "team_logo": None,
+        "team_colors": None,
+        "team_score": None,
+        "team_win_probability": None,
+        "team_timeouts": None,
+
+        "opponent_name": None,
+        "opponent_id": None,
+        "opponent_record": None,
+        "opponent_rank": None,
+        "opponent_homeaway": None,
+        "opponent_logo": None,
+        "opponent_colors": None,
+        "opponent_score": None,
+        "opponent_win_probability": None,
+        "opponent_timeouts": None,
+
+        "quarter": None,
+        "clock": None,
+        "possession": None,
+        "last_play": None,
+        "down_distance_text": None,
+
+        "outs": None,
+        "balls": None,
+        "strikes": None,
+        "on_first": None,
+        "on_second": None,
+        "on_third": None,
+
+        "team_shots_on_target": None,
+        "team_total_shots": None,
+        "opponent_shots_on_target": None,
+        "opponent_total_shots": None,
+
+        "team_sets_won": None,
+        "opponent_sets_won": None,
+
+        "last_update": None,
+        "api_message": None,
+        
+        "private_fast_refresh": False,
+    }
+
+    return new_values
